+++ src/snowflake/snowpark/dataframe_na_functions.py
@@ -9,7 +9,6 @@ from logging import getLogger
 from typing import Dict, Optional, Union
 
 import snowflake.snowpark
-from snowflake.snowpark._internal.analyzer.analyzer_utils import quote_name
 from snowflake.snowpark._internal.error_message import SnowparkClientExceptionMessages
 from snowflake.snowpark._internal.telemetry import add_api_call, adjust_api_subcalls
 from snowflake.snowpark._internal.type_utils import (
@@ -17,6 +16,7 @@ from snowflake.snowpark._internal.type_utils import (
     LiteralType,
     python_type_to_snow_type,
 )
+from snowflake.snowpark._internal.utils import quote_name
 from snowflake.snowpark.functions import iff, lit, when
 from snowflake.snowpark.types import (
     DataType,
