+++ tests/utils.py
@@ -16,11 +16,14 @@ from snowflake.connector.constants import FIELD_ID_TO_NAME
 from snowflake.snowpark import DataFrame, Row, Session
 from snowflake.snowpark._internal import utils
 from snowflake.snowpark._internal.analyzer.analyzer_utils import (
-    quote_name,
     quote_name_without_upper_casing,
 )
 from snowflake.snowpark._internal.type_utils import convert_sf_to_sp_type
-from snowflake.snowpark._internal.utils import TempObjectType, is_in_stored_procedure
+from snowflake.snowpark._internal.utils import (
+    TempObjectType,
+    is_in_stored_procedure,
+    quote_name,
+)
 from snowflake.snowpark.types import (
     ArrayType,
     BinaryType,
